spark-svc/requirements.txt
pyspark
kafka-python
pandas